+++ src/Eloquent/Builder.php
@@ -4,18 +4,14 @@ declare(strict_types=1);
 
 namespace MongoDB\Laravel\Eloquent;
 
-use Illuminate\Database\ConnectionInterface;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
-use InvalidArgumentException;
 use MongoDB\Driver\Cursor;
-use MongoDB\Laravel\Collection;
+use MongoDB\Driver\Exception\WriteException;
+use MongoDB\Laravel\Connection;
 use MongoDB\Laravel\Helpers\QueriesRelationships;
-use MongoDB\Laravel\Internal\FindAndModifyCommandSubscriber;
 use MongoDB\Laravel\Query\AggregationBuilder;
 use MongoDB\Model\BSONDocument;
-use MongoDB\Operation\FindOneAndUpdate;
 
-use function array_intersect_key;
 use function array_key_exists;
 use function array_merge;
 use function collect;
@@ -25,6 +21,7 @@ use function iterator_to_array;
 /** @method \MongoDB\Laravel\Query\Builder toBase() */
 class Builder extends EloquentBuilder
 {
+    private const DUPLICATE_KEY_ERROR = 11000;
     use QueriesRelationships;
 
     /**
@@ -201,56 +198,37 @@ class Builder extends EloquentBuilder
         return $results;
     }
 
-    /**
-     * Attempt to create the record if it does not exist with the matching attributes.
-     * If the record exists, it will be returned.
-     *
-     * @param  array $attributes The attributes to check for duplicate records
-     * @param  array $values     The attributes to insert if no matching record is found
-     */
-    public function createOrFirst(array $attributes = [], array $values = []): Model
+    public function firstOrCreate(array $attributes = [], array $values = [])
     {
-        if ($attributes === []) {
-            throw new InvalidArgumentException('You must provide attributes to check for duplicates');
+        $instance = (clone $this)->where($attributes)->first();
+        if ($instance !== null) {
+            return $instance;
         }
 
-        // Apply casting and default values to the attributes
-        // In case of duplicate key between the attributes and the values, the values have priority
-        $instance = $this->newModelInstance($values + $attributes);
-
-        /* @see \Illuminate\Database\Eloquent\Model::performInsert */
-        if ($instance->usesTimestamps()) {
-            $instance->updateTimestamps();
+        // createOrFirst is not supported in transaction.
+        if ($this->getConnection()->getSession()?->isInTransaction()) {
+            return $this->create(array_merge($attributes, $values));
         }
 
-        $values = $instance->getAttributes();
-        $attributes = array_intersect_key($attributes, $values);
+        return $this->createOrFirst($attributes, $values);
+    }
 
-        return $this->raw(function (Collection $collection) use ($attributes, $values) {
-            $listener = new FindAndModifyCommandSubscriber();
-            $collection->getManager()->addSubscriber($listener);
+    public function createOrFirst(array $attributes = [], array $values = [])
+    {
+        // The duplicate key error would abort the transaction. Using the regular firstOrCreate in that case.
+        if ($this->getConnection()->getSession()?->isInTransaction()) {
+            return $this->firstOrCreate($attributes, $values);
+        }
 
-            try {
-                $document = $collection->findOneAndUpdate(
-                    $attributes,
-                    // Before MongoDB 5.0, $setOnInsert requires a non-empty document.
-                    // This should not be an issue as $values includes the query filter.
-                    ['$setOnInsert' => (object) $values],
-                    [
-                        'upsert' => true,
-                        'returnDocument' => FindOneAndUpdate::RETURN_DOCUMENT_AFTER,
-                        'typeMap' => ['root' => 'array', 'document' => 'array'],
-                    ],
-                );
-            } finally {
-                $collection->getManager()->removeSubscriber($listener);
+        try {
+            return $this->create(array_merge($attributes, $values));
+        } catch (WriteException $e) {
+            if ($e->getCode() === self::DUPLICATE_KEY_ERROR) {
+                return $this->where($attributes)->first() ?? throw $e;
             }
 
-            $model = $this->model->newFromBuilder($document);
-            $model->wasRecentlyCreated = $listener->created;
-
-            return $model;
-        });
+            throw $e;
+        }
     }
 
     /**
@@ -276,8 +254,7 @@ class Builder extends EloquentBuilder
         return $values;
     }
 
-    /** @return ConnectionInterface */
-    public function getConnection()
+    public function getConnection(): Connection
     {
         return $this->query->getConnection();
     }